requirements.txt
cairosvg
chess
google-api-python-client
google-cloud-texttospeech
google-auth-oauthlib
moviepy
